+++ crates/ethereum/payload/src/lib.rs
@@ -10,11 +10,16 @@
 #![allow(clippy::useless_let_if_seq)]
 
 use reth_basic_payload_builder::{
-    commit_withdrawals, is_better_payload, post_block_withdrawal_requests_contract_call,
-    BuildArguments, BuildOutcome, PayloadBuilder, PayloadConfig, WithdrawalsOutcome,
+    commit_withdrawals, is_better_payload, BuildArguments, BuildOutcome, PayloadBuilder,
+    PayloadConfig, WithdrawalsOutcome,
 };
 use reth_errors::RethError;
-use reth_evm::{system_calls::pre_block_beacon_root_contract_call, ConfigureEvm};
+use reth_evm::{
+    system_calls::{
+        post_block_withdrawal_requests_contract_call, pre_block_beacon_root_contract_call,
+    },
+    ConfigureEvm,
+};
 use reth_evm_ethereum::{eip6110::parse_deposits_from_receipts, EthEvmConfig};
 use reth_execution_types::ExecutionOutcome;
 use reth_payload_builder::{
@@ -190,22 +195,24 @@ where
         }
 
         // Calculate the requests and the requests root.
-        let (requests, requests_root) =
-            if chain_spec.is_prague_active_at_timestamp(attributes.timestamp) {
-                // We do not calculate the EIP-6110 deposit requests because there are no
-                // transactions in an empty payload.
-                let withdrawal_requests = post_block_withdrawal_requests_contract_call(
-                    &mut db,
-                    &initialized_cfg,
-                    &initialized_block_env,
-                )?;
+        let (requests, requests_root) = if chain_spec
+            .is_prague_active_at_timestamp(attributes.timestamp)
+        {
+            // We do not calculate the EIP-6110 deposit requests because there are no
+            // transactions in an empty payload.
+            let withdrawal_requests = post_block_withdrawal_requests_contract_call::<EvmConfig, _>(
+                &mut db,
+                &initialized_cfg,
+                &initialized_block_env,
+            )
+            .map_err(|err| PayloadBuilderError::Internal(err.into()))?;
 
-                let requests = withdrawal_requests;
-                let requests_root = calculate_requests_root(&requests);
-                (Some(requests.into()), Some(requests_root))
-            } else {
-                (None, None)
-            };
+            let requests = withdrawal_requests;
+            let requests_root = calculate_requests_root(&requests);
+            (Some(requests.into()), Some(requests_root))
+        } else {
+            (None, None)
+        };
 
         let header = Header {
             parent_hash: parent_block.hash(),
@@ -436,11 +443,12 @@ where
     {
         let deposit_requests = parse_deposits_from_receipts(&chain_spec, receipts.iter().flatten())
             .map_err(|err| PayloadBuilderError::Internal(RethError::Execution(err.into())))?;
-        let withdrawal_requests = post_block_withdrawal_requests_contract_call(
+        let withdrawal_requests = post_block_withdrawal_requests_contract_call::<EvmConfig, _>(
             &mut db,
             &initialized_cfg,
             &initialized_block_env,
-        )?;
+        )
+        .map_err(|err| PayloadBuilderError::Internal(err.into()))?;
 
         let requests = [deposit_requests, withdrawal_requests].concat();
         let requests_root = calculate_requests_root(&requests);
