+++ crates/revm/Cargo.toml
@@ -27,7 +27,6 @@ revm.workspace = true
 
 # alloy
 alloy-eips.workspace = true
-alloy-rlp.workspace = true
 
 # common
 tracing.workspace = true
